+++ src/api/utils/export_queue.ts
@@ -25,23 +25,23 @@ type QueueOpts = {
 	callbacks?: {
 		onDone?: () => void;
 		[key: string]: Function;
-	}
+	};
 };
 
 // determines current state of a promise
 function promiseState(p: Promise<any>) {
-	const t = {}
+	const t = {};
 	return Promise.race([p, t]).then(v => (v === t) ? "pending" : "fulfilled", () => "rejected");
 }
 
 // finds first non-pending promise in a list of promises
-async function findNotPendingIndex(list: Promise<any>[]) {
+async function findNotPendingIndex(list: Array<Promise<any>>) {
 	const states = await Promise.all(list.map(p => promiseState(p)));
 	return states.findIndex(state => state !== 'pending');
 }
 
 // filters any non-pending promises out of a list of promises
-async function filterNotPending(list: Promise<any>[]) {
+async function filterNotPending(list: Array<Promise<any>>) {
 	const states = await Promise.all(list.map(p => promiseState(p)));
 	return states.reduce((acc, curr, index) => {
 		if (curr === 'pending') {
@@ -56,12 +56,12 @@ async function filterNotPending(list: Promise<any>[]) {
 // url array can contain any number of urls found during export process
 // fetching array can contain a number of fetch api returns equal to the concurrent option
 // saving array can contain a number of promisified writeFile returns equal to the concurrent option
-function exportQueue({ concurrent, handleFetch, handleResponse, fetchOpts, callbacks } : QueueOpts) {
+function exportQueue({ concurrent, handleFetch, handleResponse, fetchOpts, callbacks }: QueueOpts) {
 	const urls: URL[] = [];
-	let fetching : Promise<any>[] = [];
-	let saving : Promise<any>[] = [];
+	const fetching: Array<Promise<any>> = [];
+	let saving: Array<Promise<any>> = [];
 
-	function addToQueue(p: Promise<any>, queue: Promise<any>[]) {
+	function addToQueue(p: Promise<any>, queue: Array<Promise<any>>) {
 		const queuePromise = new Promise((res, rej) => {
 			p.then((ret?: any) => {
 				res(ret);
@@ -113,6 +113,6 @@ function exportQueue({ concurrent, handleFetch, handleResponse, fetchOpts, callb
 			callbacks[event] = fn;
 		}
 	};
-};
+}
 
 export { exportQueue, FetchOpts, FetchRet };