+++ src/core/create_compilers/RollupCompiler.ts
@@ -1,7 +1,7 @@
 import * as path from 'path';
 import color from 'kleur';
 import relative from 'require-relative';
-import { RollupError } from 'rollup/types';
+import { InputOption, RollupError } from 'rollup';
 import { CompileResult } from './interfaces';
 import RollupResult from './RollupResult';
 
@@ -13,11 +13,11 @@ export default class RollupCompiler {
 	_: Promise<any>;
 	_oninvalid: (filename: string) => void;
 	_start: number;
-	input: string;
+	input: InputOption;
 	warnings: any[];
 	errors: any[];
 	chunks: any[];
-	css_files: Array<{ id: string, code: string }>;
+	css_files: Array<{ id: string; code: string }>;
 
 	constructor(config: any) {
 		this._ = this.get_config(config);
@@ -191,7 +191,7 @@ export function handleError(err: RollupError, recover = false) {
 	}
 
 	if (err.loc) {
-		stderr(`${(err.loc.file || err.id)!} (${err.loc.line}:${err.loc.column})`);
+		stderr(`${err.loc.file || err.id} (${err.loc.line}:${err.loc.column})`);
 	} else if (err.id) {
 		stderr(err.id);
 	}