+++ src/core/create_compilers/extract_css.ts
@@ -4,7 +4,7 @@ import hash from 'string-hash';
 import * as codec from 'sourcemap-codec';
 import { PageComponent, Dirs } from '../../interfaces';
 import { CompileResult, Chunk } from './interfaces';
-import { normalize_path, posixify } from '../../utils'
+import { normalize_path, posixify } from '../../utils';
 
 const inline_sourcemap_header = 'data:application/json;charset=utf-8;base64,';
 
@@ -48,16 +48,11 @@ type SourceMap = {
 
 function get_css_from_modules(modules: string[], css_map: Map<string, string>, asset_dir: string) {
 	const parts: string[] = [];
-	const mappings: number[][][] = [];
+	const mappings: codec.SourceMapMappings = [];
 
-	const combined_map: SourceMap = {
-		version: 3,
-		file: null,
-		sources: [],
-		sourcesContent: [],
-		names: [],
-		mappings: null
-	};
+	const sources = <string[]>[]
+	const sourcesContent = <string[]>[]
+	const names = <string[]>[]
 
 	modules.forEach(module => {
 		if (!/\.css$/.test(module)) return;
@@ -71,34 +66,37 @@ function get_css_from_modules(modules: string[], css_map: Map<string, string>, a
 		if (map) {
 			const lines = codec.decode(map.mappings);
 
-			if (combined_map.sources.length > 0 || combined_map.names.length > 0) {
+			if (sources.length > 0 || names.length > 0) {
 				lines.forEach(line => {
 					line.forEach(segment => {
 						// adjust source index
-						segment[1] += combined_map.sources.length;
+						segment[1] += sources.length;
 
 						// adjust name index
-						if (segment[4]) segment[4] += combined_map.names.length;
+						if (segment[4]) segment[4] += names.length;
 					});
 				});
 			}
 
-			combined_map.sources.push(...map.sources);
-			combined_map.sourcesContent.push(...map.sourcesContent);
-			combined_map.names.push(...map.names);
+			sources.push(...map.sources);
+			sourcesContent.push(...map.sourcesContent);
+			names.push(...map.names);
 
 			mappings.push(...lines);
 		}
 	});
 
 	if (parts.length > 0) {
-		combined_map.mappings = codec.encode(mappings);
-
-		combined_map.sources = combined_map.sources.map(source => path.relative(asset_dir, source).replace(/\\/g, '/'));
-
 		return {
 			code: parts.join('\n'),
-			map: combined_map
+			map: {
+				version: 3,
+				file: null,
+				sources: sources.map(source => path.relative(asset_dir, source).replace(/\\/g, '/')),
+				sourcesContent,
+				names,
+				mappings: codec.encode(mappings)
+			},
 		};
 	}
 
@@ -113,7 +111,7 @@ export default function extract_css(
 ) {
 	const result: {
 		main: string | null;
-		chunks: Record<string, string[]>
+		chunks: Record<string, string[]>;
 	} = {
 		main: null,
 		chunks: {}
@@ -157,7 +155,7 @@ export default function extract_css(
 		}
 
 		if (sourcemap === 'inline') {
-			const base64 = Buffer.from(JSON.stringify(map), 'utf8').toString('base64')
+			const base64 = Buffer.from(JSON.stringify(map), 'utf8').toString('base64');
 			code += `\n/*# sourceMappingURL=${inline_sourcemap_header}${base64} */`;
 		}
 
@@ -272,7 +270,7 @@ export default function extract_css(
 		}
 
 		if (sourcemap === 'inline') {
-			const base64 = Buffer.from(JSON.stringify(map), 'utf8').toString('base64')
+			const base64 = Buffer.from(JSON.stringify(map), 'utf8').toString('base64');
 			code += `\n/*# sourceMappingURL=${inline_sourcemap_header}${base64} */`;
 		}