+++ src/core/create_compilers/interfaces.ts
@@ -23,18 +23,19 @@ export interface CompileResult {
 	chunks: Chunk[];
 	assets: Record<string, string>;
 	css_files: CssFile[];
+	print: () => void;
 
-	to_json: (manifest_data: ManifestData, dirs: Dirs) => BuildInfo
+	to_json: (manifest_data: ManifestData, dirs: Dirs) => BuildInfo;
 }
 
 export type BuildInfo = {
 	bundler: string;
 	shimport: string;
 	assets: Record<string, string>;
-	dependencies: Record<string, string[]>;
+	dependencies?: Record<string, string[]>;
 	legacy_assets?: Record<string, string>;
 	css: {
-		main: string | null,
-		chunks: Record<string, string[]>
-	}
+		main: string | null;
+		chunks: Record<string, string[]>;
+	};
 }