+++ src/interfaces.ts
@@ -3,10 +3,10 @@ import { CompileResult } from './core/create_compilers/interfaces';
 
 export type Route = {
 	id: string;
-	handlers: {
+	handlers: Array<{
 		type: 'page' | 'route';
 		file: string;
-	}[];
+	}>;
 	pattern: RegExp;
 	test: (url: string) => boolean;
 	exec: (url: string) => Record<string, string>;
@@ -37,7 +37,7 @@ export type Page = {
 	parts: Array<{
 		component: PageComponent;
 		params: string[];
-	}>
+	}>;
 };
 
 export type ServerRoute = {
@@ -48,9 +48,9 @@ export type ServerRoute = {
 };
 
 export type Dirs = {
-	dest: string,
-	src: string,
-	routes: string
+	dest: string;
+	src: string;
+	routes: string;
 };
 
 export type ManifestData = {
@@ -68,11 +68,19 @@ export type ReadyEvent = {
 
 export type ErrorEvent = {
 	type: string;
-	error: Error;
+	error: Error & {
+		frame?: unknown;
+		loc?: {
+			file?: string;
+			line: number;
+			column: number;
+		};
+	};
 };
 
 export type FatalEvent = {
 	message: string;
+	log?: unknown;
 };
 
 export type InvalidEvent = {
@@ -81,13 +89,13 @@ export type InvalidEvent = {
 		client: boolean;
 		server: boolean;
 		serviceworker: boolean;
-	}
+	};
 };
 
 export type BuildEvent = {
 	type: string;
-	errors: Array<{ file: string, message: string, duplicate: boolean }>;
-	warnings: Array<{ file: string, message: string, duplicate: boolean }>;
+	errors: Array<{ file: string; message: string; duplicate: boolean }>;
+	warnings: Array<{ file: string; message: string; duplicate: boolean }>;
 	duration: number;
 	result: CompileResult;
 };