+++ test/apps/errors/src/routes/trace.svelte
@@ -1,0 +1,7 @@
+<script context="module">
+	import { oops } from './_trace';
+
+	export function preload() {
+		oops();
+	}
+</script>